Remove vocabulary size from index meta Change bigram frequency to int64

+++ src/test/java/com/expleague/sensearch/donkey/plain/StatisticsBuilderTest.java
@@ -170,7 +170,7 @@ public class StatisticsBuilderTest extends SensearchTestCase {
                 bigrams.get((int) termFreq.getTermId() - 1).intValue(),
                 termFreq.getTermFrequency()));
 
-    Map<Long, Integer> freqMap =
+    Map<Long, Long> freqMap =
         bigramFrequencyList
             .stream()
                 .collect(Collectors.toMap(t -> (long) t.getTermId(), TermFrequency::getTermFrequency));
